Cleaning up project

+++ src/minecraft/fr/minecraftpp/block/BlockScenariteOre.java
@@ -2,8 +2,6 @@ package fr.minecraftpp.block;
 
 import java.util.Random;
 
-import com.google.common.base.Predicate;
-
 import fr.minecraftpp.item.ModItems;
 import net.minecraft.block.BlockOre;
 import net.minecraft.block.SoundType;
@@ -117,41 +115,41 @@ public class BlockScenariteOre extends BlockOre
 
 		for (int i = 0; i < 6; ++i)
 		{
-			double d1 = (double) ((float) pos.getX() + random.nextFloat());
-			double d2 = (double) ((float) pos.getY() + random.nextFloat());
-			double d3 = (double) ((float) pos.getZ() + random.nextFloat());
+			double d1 = pos.getX() + random.nextFloat();
+			double d2 = pos.getY() + random.nextFloat();
+			double d3 = pos.getZ() + random.nextFloat();
 
 			if (i == 0 && !world.getBlockState(pos.up()).isOpaqueCube())
 			{
-				d2 = (double) pos.getY() + 0.0625D + 1.0D;
+				d2 = pos.getY() + 0.0625D + 1.0D;
 			}
 
 			if (i == 1 && !world.getBlockState(pos.down()).isOpaqueCube())
 			{
-				d2 = (double) pos.getY() - 0.0625D;
+				d2 = pos.getY() - 0.0625D;
 			}
 
 			if (i == 2 && !world.getBlockState(pos.south()).isOpaqueCube())
 			{
-				d3 = (double) pos.getZ() + 0.0625D + 1.0D;
+				d3 = pos.getZ() + 0.0625D + 1.0D;
 			}
 
 			if (i == 3 && !world.getBlockState(pos.north()).isOpaqueCube())
 			{
-				d3 = (double) pos.getZ() - 0.0625D;
+				d3 = pos.getZ() - 0.0625D;
 			}
 
 			if (i == 4 && !world.getBlockState(pos.east()).isOpaqueCube())
 			{
-				d1 = (double) pos.getX() + 0.0625D + 1.0D;
+				d1 = pos.getX() + 0.0625D + 1.0D;
 			}
 
 			if (i == 5 && !world.getBlockState(pos.west()).isOpaqueCube())
 			{
-				d1 = (double) pos.getX() - 0.0625D;
+				d1 = pos.getX() - 0.0625D;
 			}
 
-			if (d1 < (double) pos.getX() || d1 > (double) (pos.getX() + 1) || d2 < 0.0D || d2 > (double) (pos.getY() + 1) || d3 < (double) pos.getZ() || d3 > (double) (pos.getZ() + 1))
+			if (d1 < pos.getX() || d1 > pos.getX() + 1 || d2 < 0.0D || d2 > pos.getY() + 1 || d3 < pos.getZ() || d3 > pos.getZ() + 1)
 			{
 				world.spawnParticle(EnumParticleTypes.REDSTONE, d1, d2, d3, 0.0D, 0.0D, 0.0D);
 			}
