+++ src/minecraft/net/minecraft/client/gui/MapItemRenderer.java
@@ -149,21 +149,21 @@ public class MapItemRenderer
 				if (!noOverlayRendering || mapdecoration.func_191180_f())
 				{
 					GlStateManager.pushMatrix();
-					GlStateManager.translate(0.0F + (float) mapdecoration.getX() / 2.0F + 64.0F, 0.0F + (float) mapdecoration.getY() / 2.0F + 64.0F, -0.02F);
-					GlStateManager.rotate((float) (mapdecoration.getRotation() * 360) / 16.0F, 0.0F, 0.0F, 1.0F);
+					GlStateManager.translate(0.0F + mapdecoration.getX() / 2.0F + 64.0F, 0.0F + mapdecoration.getY() / 2.0F + 64.0F, -0.02F);
+					GlStateManager.rotate(mapdecoration.getRotation() * 360 / 16.0F, 0.0F, 0.0F, 1.0F);
 					GlStateManager.scale(4.0F, 4.0F, 3.0F);
 					GlStateManager.translate(-0.125F, 0.125F, 0.0F);
 					byte b0 = mapdecoration.getType();
-					float f1 = (float) (b0 % 4 + 0) / 4.0F;
-					float f2 = (float) (b0 / 4 + 0) / 4.0F;
-					float f3 = (float) (b0 % 4 + 1) / 4.0F;
-					float f4 = (float) (b0 / 4 + 1) / 4.0F;
+					float f1 = (b0 % 4 + 0) / 4.0F;
+					float f2 = (b0 / 4 + 0) / 4.0F;
+					float f3 = (b0 % 4 + 1) / 4.0F;
+					float f4 = (b0 / 4 + 1) / 4.0F;
 					bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
 					float f5 = -0.001F;
-					bufferbuilder.pos(-1.0D, 1.0D, (double) ((float) k * -0.001F)).tex((double) f1, (double) f2).endVertex();
-					bufferbuilder.pos(1.0D, 1.0D, (double) ((float) k * -0.001F)).tex((double) f3, (double) f2).endVertex();
-					bufferbuilder.pos(1.0D, -1.0D, (double) ((float) k * -0.001F)).tex((double) f3, (double) f4).endVertex();
-					bufferbuilder.pos(-1.0D, -1.0D, (double) ((float) k * -0.001F)).tex((double) f1, (double) f4).endVertex();
+					bufferbuilder.pos(-1.0D, 1.0D, k * -0.001F).tex(f1, f2).endVertex();
+					bufferbuilder.pos(1.0D, 1.0D, k * -0.001F).tex(f3, f2).endVertex();
+					bufferbuilder.pos(1.0D, -1.0D, k * -0.001F).tex(f3, f4).endVertex();
+					bufferbuilder.pos(-1.0D, -1.0D, k * -0.001F).tex(f1, f4).endVertex();
 					tessellator.draw();
 					GlStateManager.popMatrix();
 					++k;
